Organize frontend components
