+++ README.md
@@ -57,8 +57,8 @@ recreate the tables or manually add the new unique indexes.
 The React frontend lives in the `frontend` directory. Notable reusable
 components include:
 
- - `components/DocumentList.tsx` – lists uploaded attachments with
+ - `components/ui/DocumentList.tsx` – lists uploaded attachments with
   download and delete options.
-- `components/common/Stepper.tsx` – displays progress across application
+ - `components/ui/Stepper.tsx` – displays progress across application
   steps and is used by `ApplicationLayout`.
