+++ frontend/src/pages/CallApplicationsPage.tsx
