+++ frontend/src/pages/calls/apply/Step3_Review.tsx
@@ -1,4 +1,4 @@
-import DocumentList from "../../../components/DocumentList";
+import DocumentList from "../../../components/ui/DocumentList";
 import { useApplication } from "../../../context/ApplicationProvider";
 
 export default function Step3_Review() {
